@@ -19,9 +19,9 @@
 package com.lovejjfg.powertext;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.AppCompatTextView;
 import android.util.AttributeSet;
+import androidx.annotation.Nullable;
+import androidx.appcompat.widget.AppCompatTextView;
 
 /**
  * Created by joe on 2018/1/29.
@@ -47,6 +47,4 @@ public class ClickFixedTextView extends AppCompatTextView {
     public boolean performClick() {
         return !linkClicked && super.performClick();
     }
-
-
 }
